feat: split stateful debate runner helpers
